+++ src/app/core/core.module.ts
@@ -8,6 +8,8 @@ import { TranslatorService } from "./translator/translator.service";
 import { ScrollService } from "./services/scroll.service";
 import { ColorsService } from "./services/colors.service";
 import { ACLService } from './acl/acl.service';
+import { TokenService } from './net/token/token.service';
+import { _HttpClient } from './services/http.client';
 
 @NgModule({
   imports: [
@@ -19,7 +21,9 @@ import { ACLService } from './acl/acl.service';
     TranslatorService,
     ScrollService,
     ColorsService,
-    ACLService
+    ACLService,
+    TokenService,
+    _HttpClient
   ],
   declarations: [
   ],