+++ src/app/routes/elements/sortable/sortable.component.spec.ts
@@ -1,0 +1,8 @@
+import { DemoSortableComponent } from './sortable.component';
+
+describe('Component: Sortable', () => {
+    it('should create an instance', () => {
+        const instance = new DemoSortableComponent();
+        expect(instance).toBeTruthy();
+    });
+});