+++ src/app/routes/elements/sweetalert/sweetalert.component.spec.ts
@@ -1,0 +1,8 @@
+import { SweetAlertComponent } from './sweetalert.component';
+
+describe('Component: SweetAlert', () => {
+    it('should create an instance', () => {
+        const instance = new SweetAlertComponent();
+        expect(instance).toBeTruthy();
+    });
+});