+++ src/app/routes/elements/tree-antd/tree-antd.component.spec.ts
@@ -1,0 +1,8 @@
+import { TreeAntdComponent } from './tree-antd.component';
+
+describe('Component: Tree Antd', () => {
+    it('should create an instance', () => {
+        const instance = new TreeAntdComponent();
+        expect(instance).toBeTruthy();
+    });
+});