+++ src/app/shared/shared.module.ts
@@ -12,9 +12,13 @@ import { ChartsModule } from 'ng2-charts/ng2-charts';
 import { SparklineDirective } from './directives/sparkline/sparkline.directive';
 import { MomentDatePipe } from "./pipes/moment-date.pipe";
 import { CNCurrencyPipe } from "./pipes/cn-currency.pipe";
+import { KeysPipe } from "./pipes/keys.pipe";
+import { YNPipe } from "./pipes/yn.pipe";
+import { ModalHelper } from "./helper/modal.helper";
 
 const DIRECTIVES = [SparklineDirective];
-const PIPES = [MomentDatePipe, CNCurrencyPipe];
+const PIPES = [MomentDatePipe, CNCurrencyPipe, KeysPipe, YNPipe];
+const HELPERS = [ ModalHelper ];
 
 @NgModule({
     imports: [
@@ -28,6 +32,7 @@ const PIPES = [MomentDatePipe, CNCurrencyPipe];
         ChartsModule
     ],
     declarations: [...DIRECTIVES, ...PIPES],
+    providers: [ ...HELPERS ],
     exports: [
         CommonModule,
         FormsModule,